Minor bugfix

+++ f_n2k_ngt1.cpp
@@ -195,7 +195,9 @@ void f_n2k_ngt1::handle_pgn_eng_state(PgnFieldValues * pfv,
       break;
     if(n2k_data){
       builder.Clear();
-      auto payload = NMEA2000::CreateEngineParametersRapidUpdate(builder, (NMEA2000::EngineInstance)*pfv->get_vptr<int64_t>(0), (unsigned short)*pfv->get_vptr<int64_t>(1), (unsigned short)*pfv->get_vptr<int64_t>(2), (char)*pfv->get_vptr<int64_t>(3));
+      
+      auto payload = NMEA2000::CreateEngineParametersRapidUpdate(builder,
+								 (NMEA2000::EngineInstance)*pfv->get_vptr<int64_t>(0), (unsigned short)*pfv->get_vptr<int64_t>(1), (unsigned short)*pfv->get_vptr<int32_t>(2), (char)*pfv->get_vptr<int64_t>(3));
       auto data = CreateData(builder,
 			     get_time(),
 			     NMEA2000::Payload_EngineParametersRapidUpdate,
@@ -218,13 +220,13 @@ void f_n2k_ngt1::handle_pgn_eng_state(PgnFieldValues * pfv,
 	builder.Clear();
 	auto payload =
 	  NMEA2000::CreateEngineParametersDynamic(builder, (NMEA2000::EngineInstance)*pfv->get_vptr<int64_t>(0),
-								 (unsigned short)*pfv->get_vptr<int64_t>(1), /* oil pressure */
-								 (unsigned short)*pfv->get_vptr<int64_t>(2), /* oil temperature */
-								 (unsigned short)*pfv->get_vptr<int64_t>(3), /* temperature */
+						  (unsigned short)*pfv->get_vptr<int32_t>(1), /* oil pressure */
+						  (unsigned short)*pfv->get_vptr<float64_t>(2), /* oil temperature */
+								 (unsigned short)*pfv->get_vptr<float64_t>(3), /* temperature */
 								 (unsigned short)*pfv->get_vptr<int64_t>(4), /* alternatorPotential */
 								 (short) *pfv->get_vptr<int64_t>(5), /* fuel rate */
 								 (unsigned int) *pfv->get_vptr<int64_t>(6), /* total engine hours */
-								 (unsigned short) *pfv->get_vptr<int64_t>(7), /* coolant pressure */
+								 (unsigned short) *pfv->get_vptr<int32_t>(7), /* coolant pressure */
 								 (unsigned short) *pfv->get_vptr<int64_t>(8), /* fuel pressure */
 								 (NMEA2000::EngineStatus1) *pfv->get_vptr<int64_t>(9), /* engine status 1 */
 								 (NMEA2000::EngineStatus2) *pfv->get_vptr<int64_t>(10), /* engine status 2 */
